LaTeX rendering and zip file generation now work
Still need to hook up the file stream to the transaction, but the
LaTeX is now generated, PDF files render (twice, as required for
proper formatting), and the zip archive of the test papers is
produced in an ./exams subdirectory created to receive it.

We're about ready to make history, I guess.